refactor material location
