+++ src/materials/TextureSplattingMaterial.js
@@ -1,6 +1,6 @@
 "use strict";
 
-import * as THREE from "../libs/three.module.js";
+import * as THREE from "../../libs/three.module.js";
 
 export default class TextureSplattingMaterial extends THREE.ShaderMaterial {
   /**
